refactor Chip, Image 패키지 이동 및 파일 이름 수정
