+++ core/designsystem/src/main/java/com/goalpanzi/mission_mate/core/designsystem/component/Chip.kt
@@ -1,4 +1,4 @@
-package com.goalpanzi.mission_mate.feature.onboarding.component
+package com.goalpanzi.mission_mate.core.designsystem.component
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
@@ -7,12 +7,10 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.text.TextStyle
@@ -24,7 +22,7 @@ import com.goalpanzi.mission_mate.core.designsystem.theme.ColorWhite_FFFFFFFF
 import com.goalpanzi.mission_mate.core.designsystem.theme.MissionMateTypography
 
 @Composable
-fun OutlinedTextBox(
+fun OutlinedTextChip(
     text: String,
     modifier: Modifier = Modifier,
     borderStroke: BorderStroke = BorderStroke(1.dp, ColorOrange_FFFF5732),
@@ -48,7 +46,7 @@ fun OutlinedTextBox(
 }
 
 @Composable
-fun OutlinedBox(
+fun OutlinedChip(
     modifier: Modifier = Modifier,
     backgroundColor: Color = ColorGray1_FF404249,
     borderStroke: BorderStroke = BorderStroke((0.5f).dp, ColorWhite_FFFFFFFF),
@@ -70,7 +68,7 @@ fun OutlinedBox(
 @Preview
 @Composable
 private fun PreviewOutlinedBox() {
-    OutlinedBox {
+    OutlinedChip {
         Text("test")
     }
 }
