+++ feature/board/src/main/java/com/goalpanzi/mission_mate/feature/board/component/BoardTopView.kt
@@ -27,7 +27,7 @@ import com.goalpanzi.mission_mate.core.designsystem.theme.component.NavigationTy
 import com.goalpanzi.mission_mate.feature.board.R
 import com.goalpanzi.mission_mate.feature.board.model.MissionState
 import com.goalpanzi.mission_mate.feature.board.model.UserStory
-import com.goalpanzi.mission_mate.feature.onboarding.component.StableImage
+import com.goalpanzi.mission_mate.core.designsystem.component.StableImage
 
 
 @SuppressLint("UnrememberedMutableInteractionSource")
