+++ feature/board/src/main/java/com/goalpanzi/mission_mate/feature/board/component/Piece.kt
@@ -14,18 +14,13 @@ import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -40,8 +35,8 @@ import com.goalpanzi.mission_mate.core.designsystem.theme.MissionMateTypography
 import com.goalpanzi.mission_mate.feature.board.model.BoardPiece
 import com.goalpanzi.mission_mate.feature.board.model.BoardPieceType
 import com.goalpanzi.mission_mate.feature.board.util.PieceGenerator
-import com.goalpanzi.mission_mate.feature.onboarding.component.OutlinedBox
-import com.goalpanzi.mission_mate.feature.onboarding.component.StableImage
+import com.goalpanzi.mission_mate.core.designsystem.component.OutlinedChip
+import com.goalpanzi.mission_mate.core.designsystem.component.StableImage
 
 @Composable
 fun BoxScope.Piece(
@@ -151,7 +146,7 @@ fun PieceCountChip(
     modifier: Modifier = Modifier,
     textStyle : TextStyle = MissionMateTypography.body_md_bold
 ){
-    OutlinedBox(
+    OutlinedChip(
         modifier = modifier
     ){
         Row(
